+++ dependency-manager/src/main/java/com/walmartlabs/concord/dependencymanager/DependencyManager.java
@@ -21,7 +21,6 @@ package com.walmartlabs.concord.dependencymanager;
  */
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.type.TypeFactory;
 import org.apache.maven.repository.internal.MavenRepositorySystemUtils;
 import org.eclipse.aether.*;
 import org.eclipse.aether.artifact.Artifact;
@@ -44,7 +43,10 @@ import org.eclipse.aether.util.artifact.JavaScopes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.net.URLDecoder;
 import java.nio.file.Files;
@@ -62,11 +64,9 @@ public class DependencyManager {
     private static final Logger log = LoggerFactory.getLogger(DependencyManager.class);
 
     private static final String CFG_FILE_KEY = "CONCORD_MAVEN_CFG";
-    private static final String CFG_PLUGIN_VERSION_KEY = "CONCORD_PLUGIN_VERSION_CFG";
 
     private static final String FILES_CACHE_DIR = "files";
     public static final String MAVEN_SCHEME = "mvn";
-    private static final String MVN_VERSION_PATTERN = ".*:[\\d]+(?:\\.\\d+)*(?:-\\w*)?$";
 
     private static final MavenRepository MAVEN_CENTRAL = new MavenRepository("central", "default", "https://repo.maven.apache.org/maven2/", false);
     private static final List<MavenRepository> DEFAULT_REPOS = Collections.singletonList(MAVEN_CENTRAL);
@@ -77,7 +77,6 @@ public class DependencyManager {
     private final Object mutex = new Object();
     private final RepositorySystem maven = newMavenRepositorySystem();
     private final RepositoryCache mavenCache = new DefaultRepositoryCache();
-    private final Map<String, String> pluginsVersion;
 
     public DependencyManager(Path cacheDir) throws IOException {
         this(cacheDir, readCfg());
@@ -92,15 +91,6 @@ public class DependencyManager {
 
         log.info("init -> using repositories: {}", repositories);
         this.repositories = toRemote(repositories);
-
-        String pluginsVersionCfgPath = System.getenv(CFG_PLUGIN_VERSION_KEY);
-
-        if (pluginsVersionCfgPath == null) {
-            this.pluginsVersion = Collections.emptyMap();
-        } else {
-            this.pluginsVersion = new ObjectMapper().readValue(new FileInputStream(pluginsVersionCfgPath),
-                    TypeFactory.defaultInstance().constructMapType(HashMap.class, String.class, String.class));
-        }
     }
 
     public Path getLocalCacheDir() {
@@ -142,10 +132,6 @@ public class DependencyManager {
             String scheme = item.getScheme();
             if (MAVEN_SCHEME.equalsIgnoreCase(scheme)) {
                 String id = item.getAuthority();
-
-                if (!pluginsVersion.isEmpty() && !hasVersion(id)) {
-                    id = appendVersion(id);
-                }
 
                 Artifact artifact = new DefaultArtifact(id);
 
@@ -255,7 +241,7 @@ public class DependencyManager {
         }
     }
 
-    private DefaultRepositorySystemSession newRepositorySystemSession(RepositorySystem system) throws IOException {
+    private DefaultRepositorySystemSession newRepositorySystemSession(RepositorySystem system) {
         DefaultRepositorySystemSession session = MavenRepositorySystemUtils.newSession();
         session.setCache(mavenCache);
         session.setChecksumPolicy(RepositoryPolicy.CHECKSUM_POLICY_IGNORE);
@@ -281,21 +267,6 @@ public class DependencyManager {
         });
 
         return session;
-    }
-
-    private String appendVersion(String dep) {
-        if (!isOfficialPlugin(dep)) {
-            throw new IllegalArgumentException("Unofficial plugin '" + dep + "': version is required");
-        }
-        return dep + ":" + pluginsVersion.get(dep);
-    }
-
-    private boolean isOfficialPlugin(String s) {
-        return pluginsVersion.containsKey(s);
-    }
-
-    private Boolean hasVersion(String s) {
-        return s.matches(MVN_VERSION_PATTERN);
     }
 
     private static DependencyEntity toDependency(Artifact artifact) {